Implemented district merging in single util.ordered_districts()
